g3-common: Fix parsing of signal strength responses
The stock libril has 8 bytes extra in this structure. The extra
bytes appear at the end of the structure. All the other entries look
to be identical to what libril is expecting and therefore we can
simply add 8 bytes worth of extra fields and simply ignore them until
we decide we need to see what on earth they mean.

Change-Id: I6bf8ea3a21bbe0e7658693295108d21ad59fc633